requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
yfinance>=0.2.30
pyyaml>=6.0
flask>=3.0
